requirements.txt
torch>=2.0
tokenizers>=0.14.0
transformers==4.35.0
accelerate
sympy==1.12
pebble 
timeout-decorator 
attrdict
websocket-client
